requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
reportlab==4.4.7
sqlparse==0.5.5
